+++ src/polyaburnside.py
