--- src/python/phonon_out.txt
@@ -1,5 +1,0 @@
-[0], [1, 2]
-[[-1, 1], [0, 2]]
-[[-1, 1], [1, 2]]
-[[-1, 1], [4, 2]]
-3
